changing df to df0 to prevent error
In the section `Applying User-Defined Function`(http://spark.apache.org/docs/2.0.0/sparkr.html#run-a-given-function-on-a-large-dataset-using-dapply-or-dapplycollect), dapply and dapplyCollect are using the same SparkDataFrame `df` in which a new column `waiting_secs` was added in the section `Operating on Columns`. Now, if a user is sequentially executing the code, it will throw an error when schema is applied on `df` within dapply function to create new SparkDataFrame `df1`.